// migrations/snapshots/06c3ae25958e370a3f11c04274d6be5c87bb8ef16703309fd6f3e0bac9f4be1e/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly FeedItem: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly sessionId: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly position: CodecTypes['pg/int4@1']['output'];
      readonly isOrg: CodecTypes['pg/bool@1']['output'];
      readonly reason: CodecTypes['pg/text@1']['output'];
    };
    readonly FeedSession: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly highWaterMark: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly itemCount: CodecTypes['pg/int4@1']['output'];
      readonly openedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Organization: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly mark: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly PollOption: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly votes: CodecTypes['pg/int4@1']['output'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'] | null;
      readonly body: CodecTypes['pg/text@1']['output'] | null;
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'] | null;
      readonly topicId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'] | null;
      readonly orgId: CodecTypes['pg/text@1']['output'] | null;
      readonly art: CodecTypes['pg/text@1']['output'] | null;
      readonly cap: CodecTypes['pg/text@1']['output'] | null;
      readonly sgi: CodecTypes['pg/bool@1']['output'];
      readonly src: CodecTypes['pg/text@1']['output'] | null;
      readonly why: CodecTypes['pg/text@1']['output'] | null;
      readonly likesCount: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly PostSeen: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly seenAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Reaction: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Reel: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly subtitle: CodecTypes['pg/text@1']['output'];
      readonly duration: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Save: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Topic: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly TopicFollow: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly topicId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly avatar: CodecTypes['pg/text@1']['output'] | null;
      readonly timezone: CodecTypes['pg/text@1']['output'] | null;
      readonly deliveryWindows: ReadonlyArray<CodecTypes['pg/int4@1']['output']>;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly FeedItem: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly sessionId: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly position: CodecTypes['pg/int4@1']['input'];
      readonly isOrg: CodecTypes['pg/bool@1']['input'];
      readonly reason: CodecTypes['pg/text@1']['input'];
    };
    readonly FeedSession: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly highWaterMark: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly itemCount: CodecTypes['pg/int4@1']['input'];
      readonly openedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Organization: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly mark: CodecTypes['pg/text@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly PollOption: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly votes: CodecTypes['pg/int4@1']['input'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'] | null;
      readonly body: CodecTypes['pg/text@1']['input'] | null;
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'] | null;
      readonly topicId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'] | null;
      readonly orgId: CodecTypes['pg/text@1']['input'] | null;
      readonly art: CodecTypes['pg/text@1']['input'] | null;
      readonly cap: CodecTypes['pg/text@1']['input'] | null;
      readonly sgi: CodecTypes['pg/bool@1']['input'];
      readonly src: CodecTypes['pg/text@1']['input'] | null;
      readonly why: CodecTypes['pg/text@1']['input'] | null;
      readonly likesCount: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly PostSeen: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly seenAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Reaction: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Reel: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly subtitle: CodecTypes['pg/text@1']['input'];
      readonly duration: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Save: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Topic: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly TopicFollow: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly topicId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly avatar: CodecTypes['pg/text@1']['input'] | null;
      readonly timezone: CodecTypes['pg/text@1']['input'] | null;
      readonly deliveryWindows: ReadonlyArray<CodecTypes['pg/int4@1']['input']>;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;