// migrations/snapshots/5e64a70263b8a893329a9d443f3edb845d5fd805544a4a53dc7b6a4d904acdda/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Organization: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly mark: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly PollOption: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly votes: CodecTypes['pg/int4@1']['output'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'] | null;
      readonly body: CodecTypes['pg/text@1']['output'] | null;
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'] | null;
      readonly topicId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'] | null;
      readonly orgId: CodecTypes['pg/text@1']['output'] | null;
      readonly art: CodecTypes['pg/text@1']['output'] | null;
      readonly cap: CodecTypes['pg/text@1']['output'] | null;
      readonly sgi: CodecTypes['pg/bool@1']['output'];
      readonly src: CodecTypes['pg/text@1']['output'] | null;
      readonly why: CodecTypes['pg/text@1']['output'] | null;
      readonly likesCount: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Reaction: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Reel: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly subtitle: CodecTypes['pg/text@1']['output'];
      readonly duration: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Save: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Topic: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly avatar: CodecTypes['pg/text@1']['output'] | null;
      readonly timezone: CodecTypes['pg/text@1']['output'] | null;
      readonly deliveryWindows: ReadonlyArray<CodecTypes['pg/int4@1']['output']>;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Organization: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly mark: CodecTypes['pg/text@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly PollOption: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly votes: CodecTypes['pg/int4@1']['input'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'] | null;
      readonly body: CodecTypes['pg/text@1']['input'] | null;
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'] | null;
      readonly topicId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'] | null;
      readonly orgId: CodecTypes['pg/text@1']['input'] | null;
      readonly art: CodecTypes['pg/text@1']['input'] | null;
      readonly cap: CodecTypes['pg/text@1']['input'] | null;
      readonly sgi: CodecTypes['pg/bool@1']['input'];
      readonly src: CodecTypes['pg/text@1']['input'] | null;
      readonly why: CodecTypes['pg/text@1']['input'] | null;
      readonly likesCount: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Reaction: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Reel: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly subtitle: CodecTypes['pg/text@1']['input'];
      readonly duration: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Save: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Topic: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly avatar: CodecTypes['pg/text@1']['input'] | null;
      readonly timezone: CodecTypes['pg/text@1']['input'] | null;
      readonly deliveryWindows: ReadonlyArray<CodecTypes['pg/int4@1']['input']>;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly organization: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly mark: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly pollOption: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly votes: CodecTypes['pg/int4@1']['output'];
    };
    readonly post: {
      readonly art: CodecTypes['pg/text@1']['output'] | null;
      readonly body: CodecTypes['pg/text@1']['output'] | null;
      readonly cap: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'] | null;
      readonly likesCount: CodecTypes['pg/int4@1']['output'];
      readonly orgId: CodecTypes['pg/text@1']['output'] | null;
      readonly sgi: CodecTypes['pg/bool@1']['output'];
      readonly src: CodecTypes['pg/text@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'] | null;
      readonly topicId: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'] | null;
      readonly why: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly reaction: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly reel: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly duration: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly subtitle: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
    readonly save: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly topic: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly user: {
      readonly avatar: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly deliveryWindows: ReadonlyArray<CodecTypes['pg/int4@1']['output']>;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly timezone: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;